Implement new db interface
