+++ .gitignore
@@ -3,3 +3,4 @@ test-pull
 test-push
 test-server
 lit
+test
